requirements.txt
pathway==0.12.0
openai>=1.0.0
sentence-transformers>=2.2.0
torch
streamlit>=1.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic
pandas
transformers # Explicitly add transformers if pinning (optional for now)